# Mac: Clean up kext perf tracing patterns, and a significant perf fix