… ProjFS.Mac/PrjFSKext/PrjFSKext/PerformanceTracing.hpp
… ProjFS.Mac/PrjFSKext/PrjFSKext/PrjFSLogUserClient.cpp
… ProjFS.Mac/PrjFSKext/PrjFSKext/VirtualizationRoots.cpp
… ProjFS.Mac/PrjFSKext/PrjFSKext/VirtualizationRoots.hpp
… ProjFS.Mac/PrjFSKext/public/PrjFSCommon.h
… ProjFS.Mac/PrjFSLib/prjfs-log/kext-perf-tracing.cpp