Commit 1:

I reworked the perf tracing patterns to give us a hierarchical breakdown of where the time is going in the kext, similar to the function call tree produced by the VS perf profiler. Each node in the report represents the inclusive time spent in a pseudo function call (it's pseudo because in some cases it's just a block of code, not necessarily an actual function call, but logically you can interpret it the same way).

With those changes, I got the following report after running the Adium build on a fresh MirrorProvider clone:
```
   Counter                             [ Samples  ][Total time (ns)][Mean (ns)   ][Stddev (ns) ][Min (ns)][Max (ns)  ]
----------------------------------------------------------------------------------------------------------------------
 0 HandleVnodeOperation                [  12707740][    68112342790][        5360][      176625][     198][ 170716142]
 1  |--ShouldHandleVnodeOpEvent        [  12707740][    53659192222][        4223][       68021][      63][ 164240990]
 2  |  |--VnodeIsOnAllowedFilesystem   [  12707740][      393173955][          31][        5781][      17][  14733583]
 3  |  |--ShouldIgnoreVnodeType        [  12701685][      250090086][          20][        1039][      14][   2906761]
 4  |  |  |--Ignored                   [     15387]
 5  |  |--TryReadVNodeFileFlags        [  12686298][     1368471236][         108][        5451][      63][  17896845]
 6  |  |--NotInAnyRoot                 [   9064092]
 7  |  |--IsFileSystemCrawler          [    132211][        4928745][          37][          87][      16][     19530]
 8  |  |  |--Denied                    [      9414]
 9  |  |--FindVirtualizationRoot       [   3612792][    48182699370][       13337][       57320][      88][  53361815]
10  |  | |--TemporaryDirectory         [         0]
11  |  | |--NoRootFound                [         0]
12  |  | |--ProviderOffline            [         1]
13  |  | |--OriginatedByProvider       [   1104872]
14  |  |--IsHandledEvent               [   2507919]
15  |--RaisePreDeleteEvent             [       702][      453812446][      646456][     1296294][   82639][  17185477]
16  |--RaiseEnumerateDirectoryEvent    [       697][     5391452218][     7735226][    20086446][  613829][ 170692499]
17  |--RaiseRecursivelyEnumerateEvent  [         0]
18  |--RaiseHydrateFileEvent           [      5509][     4941968414][      897072][     1259833][  279390][  78393639]
```

Summary of interesting findings:
* 68.1s spent handling vnode callbacks
* 48.2s spent finding virtualization roots
* 10.8s spent blocking on user mode for enumeration/hydration/predelete events
* Adium build wall time was 00:02:02

The biggest thing that jumped out here is that we spent 48s to find a virtualization root 3.6M times, but only needed to send a message to the provider ~7K times. In other words, >99% of those expensive lookups were wasted. Which led to the subsequent change:

Commit 2:

Move the lookup of a virtualization root to happen as late as possible, right before sending a message to the provider. This allows the other cheap checks to filter out the majority of vnode operations before we ever do this expensive check. The perf report from running the same Adium build on a fresh MirrorProvider clone:

```
   Counter                             [ Samples  ][Total time (ns)][Mean (ns)   ][Stddev (ns) ][Min (ns)][Max (ns)  ]
----------------------------------------------------------------------------------------------------------------------
 0 HandleVnodeOperation                [  13123630][    16878629611][        1286][      126560][     214][ 265931393]
 1  |--ShouldHandleVnodeOpEvent        [  13123630][     4612836744][         351][       33280][      61][  68108690]
 2  |  |--VnodeIsOnAllowedFilesystem   [  13123630][      360144701][          27][         612][      17][   1645261]
 3  |  |--ShouldIgnoreVnodeType        [  13117668][      270749407][          21][        5667][      14][  19861202]
 4  |  |  |--Ignored                   [     15404]
 5  |  |--TryReadVNodeFileFlags        [  13102264][     1440586257][         110][        6389][      63][  16794546]
 6  |  |--NotInAnyRoot                 [   9241813]
 7  |  |--IsFileSystemCrawler          [    134218][        4814707][          36][         175][      16][     53968]
 8  |  |  |--Denied                    [     10233]
 9  |--TryGetVirtualizationRoot        [    106131][     1358659513][       12802][       15098][     144][   4298306]
10  | |--TemporaryDirectory            [         0]
11  | |--NoRootFound                   [         0]
12  | |--ProviderOffline               [         1]
13  | |--OriginatedByProvider          [     99210]
14  |--RaisePreDeleteEvent             [       702][      351894570][      501274][     1198590][   64217][  12221165]
15  |--RaiseEnumerateDirectoryEvent    [       708][     3633374439][     5131885][    15552841][  519950][ 265924745]
16  |--RaiseRecursivelyEnumerateEvent  [         0]
17  |--RaiseHydrateFileEvent           [      5510][     3248656033][      589593][      479045][  114055][  16512466]
```

Summary:
* 16.9s spent handling vnode callbacks. 75% improvement.
* 1.4s spent finding virtualization roots. 97% improvement.
* 7.2s spent blocking on user mode for enumeration/hydration/predelete events. 33% improvement.
* Adium build wall time was 00:01:36. 27% improvement.

I was expecting the 97% drop in finding virtualization roots, but the 33% improvement on handling events was a nice surprise. It makes sense that it happened, because those events cause the provider to do reentrant IO, which has to go through all of our checks before early-exiting. The more events we can discard cheaply, the less expensive that is.

Commit 3:
Added file op perf samples:
```
   Counter                             [ Samples  ][Total time (ns)][Mean (ns)   ][Stddev (ns) ][Min (ns)][Max (ns)  ]
----------------------------------------------------------------------------------------------------------------------
 0 HandleVnodeOperation                [  13163989][    18981657304][        1442][      118459][     227][ 184152944]
 1  |--GetPath                         [  13163989][     3186550902][         242][        6044][      64][  20061121]
 2  |--ShouldHandleVnodeOpEvent        [  13163989][     4354062234][         331][        7831][      60][  18586350]
 3  |  |--VnodeIsOnAllowedFilesystem   [  13163989][      374080971][          28][         788][      17][   1503914]
 4  |  |--ShouldIgnoreVnodeType        [  13157552][      263898411][          20][         497][      14][   1484400]
 5  |  |  |--Ignored                   [     15384]
 6  |  |--TryReadVNodeFileFlags        [  13142168][     1557629604][         119][        5516][      63][  18585547]
 7  |  |  |--NotInAnyRoot              [   9835331]
 8  |  |--IsFileSystemCrawler          [    129679][        4427817][          34][         143][      16][     36658]
 9  |     |--Denied                    [      6860]
10  |--TryGetVirtualizationRoot        [    105129][     1399170861][       13309][      274580][     195][  88914298]
11  | |--TemporaryDirectory            [         0]
12  | |--NoRootFound                   [         0]
13  | |--ProviderOffline               [         1]
14  | |--CompareProviderPid            [    105128][        5031724][          48][         102][      31][     17361]
15  |    |--OriginatedByProvider       [     98214]
16  |--RaisePreDeleteEvent             [       702][      368488951][      524913][     1380837][   57948][  14399688]
17  |--RaiseEnumerateDirectoryEvent    [       697][     3595527295][     5158576][    14000024][  572801][ 184145613]
18  |--RaiseRecursivelyEnumerateEvent  [         0]
19  |--RaiseHydrateFileEvent           [      5515][     4607823259][      835507][     1598198][  198367][  96513269]
20 HandleFileOpOperation               [    320053][    10992749994][       34347][      284229][     119][  47542333]
21  |--ShouldHandleFileOpEvent         [     86892][     3686937132][       42431][      135410][      17][  14662369]
22  |  |--FindVirtualizationRoot       [     68160][     3672097630][       53875][      150802][    3054][  14661914]
23  |  |  |--NoRootFound               [     46296]
24  |  |--CompareProviderPid           [     21864][        1893561][          87][         124][      43][     17749]
25  |     |--OriginatedByProvider      [     11177]
26  |--RaiseRenamedEvent               [       704][      473367621][      672397][     1614217][  144496][  26175324]
27  |--RaiseHardLinkCreatedEvent       [         0]
28  |--RaiseFileModifiedEvent          [         5][         846182][      169236][      127364][   76061][    391567]
29  |--RaiseFileCreatedEvent           [      9978][     6645490642][      666014][     1306064][  135486][  47508696]
```

Commit 4:
Added the ability to sample a subset of events, set at 1% for this commit. The relative times spent on various events did prove to be representative, though I also saw that it did not affect the build wall time in any meaningful way, so this could just be unnecessary complexity.

Commit 5:
Added perf samples inside VirtualizationRoot_FindForVnode, which shows a very interesting result. In the vnode path, we spent 11894ns per call, and for fileop we spend 51954ns per call, more than 4x more expensive. And then it turns out that of the 69K times we called VirtualizationRoot_FindForVnode, we found no root at all and aborted 48K times, so about 70% of that work was wasted. Though in actuality the waste was even more than 70%, because the calls that find no root also have to do more work to determine that there is no root - the evidence for that is that in the vnode path, each call to VirtualizationRoot_FindForVnode on average runs through 5 loop iterations, and in the fileop path it runs through almost 12 loop iterations.

Sample report:
```
   Counter                             [ Samples  ][Total time (ns)][Mean (ns)   ][Stddev (ns) ][Min (ns)][Max (ns)  ]
----------------------------------------------------------------------------------------------------------------------
 0 HandleVnodeOperation                [  12075273][    17782543772][        1473][      118093][     230][ 167150514]
 1  |--GetPath                         [  12075273][     2769819839][         229][        7982][      64][  19619394]
 2  |--ShouldHandleVnodeOpEvent        [  12075273][     4054993505][         336][        9341][      64][  16618261]
 3  |  |--VnodeIsOnAllowedFilesystem   [  12075273][      335893500][          28][         472][      17][   1073367]
 4  |  |--ShouldIgnoreVnodeType        [  12069381][      255930151][          21][        2450][      14][   6879182]
 5  |  |  |--Ignored                   [     15375]
 6  |  |--TryReadVNodeFileFlags        [  12054006][     1411611349][         117][        4400][      63][   9068230]
 7  |  |  |--NotInAnyRoot              [   8657823]
 8  |  |--IsFileSystemCrawler          [    128708][        4847459][          38][         941][      16][    335447]
 9  |     |--Denied                    [      5910]
10  |--TryGetVirtualizationRoot        [    105108][     1275482408][       12135][       10662][     284][   1846149]
11  | |--FindForVnode                  [    105108][     1250134926][       11894][       10582][     114][   1845678]
12  |    |--inner_loop_iterations      [    540665][     1224519773][        2265][        3772][      32][   1798062]
13  | |--TemporaryDirectory            [         0]
14  | |--NoRootFound                   [         0]
15  | |--ProviderOffline               [         1]
16  | |--CompareProviderPid            [    105107][        5495839][          52][         254][      32][     73417]
17  |    |--OriginatedByProvider       [     98200]
18  |--RaisePreDeleteEvent             [       702][      394358492][      561764][     1424549][   53836][  23332103]
19  |--RaiseEnumerateDirectoryEvent    [       697][     3531676384][     5066968][    13498210][  475475][ 167145767]
20  |--RaiseRecursivelyEnumerateEvent  [         0]
21  |--RaiseHydrateFileEvent           [      5508][     4430083131][      804300][     1715637][  277148][ 107526201]
22 HandleFileOpOperation               [    303900][    11441961970][       37650][      327558][     120][  48615847]
23  |--ShouldHandleFileOpEvent         [     88186][     3643432750][       41315][       54599][      16][   8381901]
24  |  |--FindVirtualizationRoot       [     69540][     3627646064][       52166][       56552][    2834][   8381232]
25  |  |  |--FindForVnode              [     69540][     3612912798][       51954][       56524][    2731][   8380158]
26  |  |  |  |--inner_loop_iterations  [    820954][     3570870463][        4350][       13843][      33][   8230763]
27  |  |  |--NoRootFound               [     47676]
28  |  |--CompareProviderPid           [     21864][        2036180][          93][         192][      44][     17484]
29  |     |--OriginatedByProvider      [     11177]
30  |--RaiseRenamedEvent               [       704][      507238234][      720509][     1909627][  164788][  24991258]
31  |--RaiseHardLinkCreatedEvent       [         0]
32  |--RaiseFileModifiedEvent          [         5][        3778981][      755796][      870215][  180864][   2278100]
33  |--RaiseFileCreatedEvent           [      9978][     7075090178][      709069][     1539321][  144888][  48569136]
```

Report produced by final commit:
```
   Counter                             [ Samples  ][Total time (ns)][Mean (ns) ][Min (ns)  ][Max (ns)  ]
--------------------------------------------------------------------------------------------------------
 0 HandleVnodeOperation                [    129898][      208832136][      1607][       373][  35959438]
 1  |--GetPath                         [    129898][       31871509][       245][        68][    377138]
 2  |--ShouldHandleVnodeOpEvent        [    129898][       48940604][       376][        79][   1311048]
 3  |  |--VnodeIsOnAllowedFilesystem   [    129898][        5502337][        42][        18][     17406]
 4  |  |--ShouldIgnoreVnodeType        [    129845][        3557683][        27][        14][     12791]
 5  |  |  |--Ignored                   [       142]
 6  |  |--TryReadVNodeFileFlags        [    129703][       17594369][       135][        65][    203185]
 7  |  |  |--NotInAnyRoot              [     92132]
 8  |  |--IsFileSystemCrawler          [      1348][          79204][        58][        17][       670]
 9  |     |--Denied                    [        74]
10  |--TryGetVirtualizationRoot        [      1093][       13557489][     12403][      2335][    957254]
11  |  |--FindForVnode                 [      1093][       13245022][     12118][      2112][    956743]
12  |  |  |--inner_loop_iterations     [      5617][       12951534][      2305][        36][    929036]
13  |  |--TemporaryDirectory           [         0]
14  |  |--NoRootFound                  [         0]
15  |  |--ProviderOffline              [         0]
16  |  |--CompareProviderPid           [      1093][          82629][        75][        36][     12724]
17  |     |--OriginatedByProvider      [      1027]
18  |--RaisePreDeleteEvent             [        12][        3174673][    264556][     69587][    957182]
19  |--RaiseEnumerateDirectoryEvent    [         8][       54204190][   6775523][    615127][  35945986]
20  |--RaiseRecursivelyEnumerateEvent  [         0]
21  |--RaiseHydrateFileEvent           [        46][       37810336][    821963][    357905][   1902284]
22 HandleFileOpOperation               [      3108][       86564904][     27852][       133][   5427074]
23  |--ShouldHandleFileOpEvent         [       849][       34283750][     40381][        19][   1107132]
24  |  |--FindVirtualizationRoot       [       663][       34138354][     51490][      5376][   1106944]
25  |  |  |--FindForVnode              [       663][       33989074][     51265][      5257][   1106691]
26  |  |  |  |--inner_loop_iterations  [      7481][       33596391][      4490][        47][   1033460]
27  |  |  |--NoRootFound               [       466]
28  |  |--CompareProviderPid           [       197][          18984][        96][        53][       217]
29  |     |--OriginatedByProvider      [       102]
30  |--RaiseRenamedEvent               [         5][        1459451][    291890][    189457][    383921]
31  |--RaiseHardLinkCreatedEvent       [         0]
32  |--RaiseFileModifiedEvent          [         0]
33  |--RaiseFileCreatedEvent           [        90][       49165660][    546285][    132274][   5396600]
```